Add a new Term constructor: Anything
This is for use in Rules for terms of a literal that are allowed to
take on any value.  This is syntactic sugar for allocating a fresh
logic variable (in fact, that is what happens internally).

This also includes a port of the test case included with the clojure
implementation of datalog.  Only the first query is included so far.