Massive allocation improvement.
Store relation data in a list (with a side hash set for uniqueness).
This is more up-front storage but it means that we can traverse the
list cheaply instead of using HashSet foldr/foldl, which allocate just
to do a traversal.

This brought the anyMatch function down to zero allocations.  The
overall profile is much better now, though that function must still do
a lot of work.