Add a unique identifier to each conditional clause literal.
Since these literals contain actual Haskell functions, comparing them
is difficult.  This change gives each one a unique id that can be used
in comparisons later.  This was necessary because rules must be looked
up in a map during query planning and, without a way for conditional
clauses to compare as equal, that lookup cannot succeed.

This change does not break the API unless you used type annotations on
fragments of rules.